addNode as shared func
